process_photos/requirements.txt
boto3==1.26.137
Pillow